backend/requirements.txt
flask
flask_cors
flask_restful
flask_jwt_extended
gunicorn
motor
pymongo
python-dotenv
phonenumbers
passlib[bcrypt]
PyJWT
python-jose[cryptography]
pydantic
shortuuid
redis
